requirements.txt
-e git+https://github.com/aristanetworks/arcomm.git#egg=arcomm
Jinja2>=2.10
mistune>=0.8.1
pytest>=3.2.5
PyYAML>=3.12
requests>=2.18.4
ipython>=6.2.1
